# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Quizzy API",
        "version": "0.1"
    },
    "paths": {
        "/api/questions": {
            "get": {
                "summary": "Get All Questions",
                "description": "Get list of all questions",
                "operationId": "findPetsByTags",
                "responses": {
                    "200": {
                        "description": "successful operation"
                    },
                    "400": {
                        "description": "Invalid tag value"
                    }
                },
                "security": [
                    {
                        "petstore_auth": [
                            "write:pets",
                            "read:pets"
                        ]
                    }
                ]
            },
            "post": {
                "summary": "Create a question",
                "description": "Create a question",
                "operationId": "findPetsByTags",
                "requestBody": {
                    "description": "Data required for creating the question",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "type": "array",
                                "items": {
                                    "properties": {
                                        "content": {
                                            "type": "string",
                                            "example": "Which popular video game franchise has released games with the subtitles World At War and Black Ops?"
                                        },
                                        "typeId": {
                                            "type": "integer",
                                            "example": "1"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "successful operation"
                    },
                    "400": {
                        "description": "Invalid tag value"
                    }
                },
                "security": [
                    {
                        "petstore_auth": [
                            "write:pets",
                            "read:pets"
                        ]
                    }
                ]
            }
        }
    }
}